Add ArtifactBuilder to add abstraction between service and image
Abstract away the docker-compose build into a class called ArtifactBuilder.
Allows us to plugin future artifact building engine, as long as they map
the Docker image created to an intermediate artifact ID (in this case,
the artifact ID is the docker-compose service). We can then map from that
artifact ID to all the repos and tags we need.

This allows us to contain all configuration for the aritfact building in
that class, and allows us to use different tools for artifact building in the
future (e.g. Skaffold, garden.io, Tilt in the Kubernetes world, or just
a config file that maps from an artifact ID to a Dockerfile.)